refactor: [박지영] donate 도메인 리팩토링 donate → board.donate [BYL-FRONT-WEB-UI-16]
기존 donate 도메인 리팩토링
- [x] components.donate -> components.boards.donate
- [x] views.donate -> views.boards.donate

- [x] components 파일은 모두 DonateChoice`Form`.vue 이런식으로 바꿈

- [x] views 파일은 모두 DonateChoice`View`.vue 이런식으로 바꿈
