+++ byl_web/src/views/boards/donate/DonateChoiceView.vue
@@ -1,11 +1,11 @@
 <template>
     <v-container>
-    <donate-choice-form/>
-  </v-container>
+      <donate-choice-form/>
+    </v-container>
 </template>
 
 <script>
-  import DonateChoiceForm from "@/components/donate/DonateChoiceForm";
+  import DonateChoiceForm from "@/components/boards/donate/DonateChoiceForm";
   
   export default {
     name: "DonateChoicePage",
