+++ byl_web/src/views/boards/donate/DonateMailView.vue
@@ -1,16 +1,17 @@
 <template>
-      <v-container>
+    <v-container>
       <donate-mail-form/>
     </v-container>
 </template>
 
 <script>
-    import DonateMailForm from "@/components/donate/DonateMailForm";
+    import DonateMailForm from "@/components/boards/donate/DonateMailForm";
     
     export default {
       name: "DonateMailPage",
       components: {
         DonateMailForm
+        
       }
     }
 </script>
