+++ byl_web/src/views/boards/donate/DonateSearchView.vue
@@ -1,11 +1,11 @@
 <template>
-    <v-container>
+  <v-container>
     <donate-search-form/>
   </v-container>
 </template>
 
 <script>
-  import DonateSearchForm from "@/components/donate/DonateSearchForm";
+  import DonateSearchForm from "@/components/boards/donate/DonateSearchForm";
   
   export default {
     name: "DonateSearchPage",
